+++ lib/src/enums/enums.dart
@@ -1,6 +1,7 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // Barrel file for all generated enums
 
+export 'material_certifications_enum.enum.g.dart';
 export 'material_class_enum.enum.g.dart';
 export 'material_type_enum.enum.g.dart';
 export 'tag_categories_enum.enum.g.dart';